Build time fix for index data types when loading sequences from OMA DB. Bump to v2.0.5

+++ omamer/database.py
@@ -1549,8 +1549,8 @@ class DatabaseFromOMABrowser(DatabaseFromOMA):
 
                     # sequence
                     oma_seq_off = rr["SeqBufferOffset"]
-                    seq_len = rr["SeqBufferLength"] - 1
-                    seq = oma_seq_buffer[oma_seq_off : oma_seq_off + seq_len + 1]
+                    seq_len = rr["SeqBufferLength"] - np.uint64(1)
+                    seq = oma_seq_buffer[oma_seq_off : oma_seq_off + seq_len + np.uint64(1)]
                     seq_buffs.append(seq)
 
                     # store protein row
